model_training/requirements.txt
tensorflow
Flask
requests
azure-storage-blob
pymongo
